+++ tests/Unit/DependencyInjection/ConfigurationTest.php
@@ -57,6 +57,7 @@ class ConfigurationTest extends TestCase
         $this->assertStringContainsString('ignore_access_denied_listener:', $content);
         $this->assertStringContainsString('sub_request_access_denied_listener:', $content);
         $this->assertStringContainsString('before_send_handler:', $content);
+        $this->assertStringContainsString('before_send_transaction_handler:', $content);
         $this->assertStringContainsString('uptime_bot_listener:', $content);
         $this->assertStringContainsString('error_reporter:', $content);
         $this->assertStringContainsString('dbal_exception_reporter:', $content);